add article functionality
